+++ NEWS.md
@@ -8,6 +8,7 @@ None
 
 * Created new vignettes "Types of VIMs" and "Using precomputed regression function estimates in `vimp`"
 * Updated main vignette to only use `run_regression = TRUE` for simplicity
+* Added argument `verbose` to `sp_vim`; if `TRUE`, messages are printed throughout fitting that display progress and `verbose` is passed to `SuperLearner`
 
 # vimp 2.1.0